1 - Initial Work
This is a simulator of a board game ostromo. It is supposed to give an
indicator on how difficult the game is. It ignores the multiplayer
aspect of the game and just acts as if there is one player. Can that
player beat the game with the resources given?

The AI for the game is fairly naive (it just uses some heuristics and
doesn't do any tree searching).

I'm using it to tune the resources given to players in the multi-player
board game. If the simulator always wins, then it is too easy, if it always
loses then it is too hard.

I'll improve the simulator at some point to play at a higher level of
play. Then try and give easy/hard options for the board game.

This could be the start of an on-line version of the game. Create an issue
if you want something like this and we'll see how we can do this.